Added accept click listener. Pass type to detail page.
For now, I passed the type as 'data' when each item in the feed_recycler is clicked. This is used to set the correct type in the detail page. We will need to later probably pass an entire database entry on the intent and pull the data from there instead.